mypy: fix errors in remaining misc root-level modules
Fixes the last batch of mypy errors across strain_library.py,
image_manager.py, view_model_builder.py's CacheManager dependency,
notifications/, storage_manager.py, bayesian_evaluator.py,
trend_analyzer.py, tank_monitor.py, substrate_tracker.py,
notification_manager.py, and utils.py. `uv run --no-sync mypy
custom_components/growspace_manager` now reports zero errors.

Closes #611

+++ custom_components/growspace_manager/substrate_tracker.py
@@ -363,7 +363,7 @@ class SubstrateTracker:
         events = self.get_incycle_drybacks_today(reference_ts)
         if not events:
             return None
-        return round(sum(e["dryback"] for e in events) / len(events), 4)
+        return float(round(sum(e["dryback"] for e in events) / len(events), 4))
 
     def get_shot_count_today(self, reference_ts: str | None = None) -> int:
         """Return the number of P2 in-cycle dryback shot-pairs recorded today.
